* Fix the bug related with loading session.     Now coordinates stores in the .vdb file as integer value
git-svn-id: https://visiual-db.svn.sourceforge.net/svnroot/visiual-db/trunk@184 a9af4176-ed96-4626-b17c-dfd41ec4b82a

+++ src/gui/TableItem.cpp
@@ -579,8 +579,8 @@ TableItem::toXml(QDomDocument &ipDoc)
     element.setAttribute("schema", schemaName());
     element.setAttribute("name", tableName());
     QPointF point = mapToScene(QPointF(x(), y()));
-    element.setAttribute("x", point.x());
-    element.setAttribute("y", point.y());
+    element.setAttribute("x", (int)point.x());
+    element.setAttribute("y", (int)point.y());
     element.setAttribute("width", width());
     element.setAttribute("height", height());
     element.setAttribute("red", itemColor().red());
